Move storage manager to separate file #45
