+++ src/utils.ts
@@ -1,4 +1,4 @@
-import { Message, storageKey, StorageStructure } from './const';
+import { Message } from './const';
 import { type ClassValue, clsx } from 'clsx';
 import { twMerge } from 'tailwind-merge';
 
@@ -23,29 +23,3 @@ export function sendMessage(message: Message) {
     chrome.runtime.sendMessage(message);
   } catch (e) {}
 }
-
-class StorageManager {
-  private _storage: chrome.storage.SyncStorageArea;
-
-  constructor() {
-    this._storage = chrome.storage.sync;
-  }
-
-  get<K extends storageKey>(key: K): Promise<StorageStructure[K]> {
-    return new Promise((resolve, reject) => {
-      this._storage.get([key], (result) => {
-        resolve(result[key]);
-      });
-    });
-  }
-
-  set<K extends storageKey>(key: K, value: StorageStructure[K]): Promise<void> {
-    return new Promise((resolve, reject) => {
-      this._storage.set({ [key]: value }, () => {
-        resolve();
-      });
-    });
-  }
-}
-
-export const storage = new StorageManager();
